Split out connector and add connector options
